move test cases for implicit relations to separate file
