+++ libs/datamodel/core/tests/reformat/mod.rs
@@ -1,1 +1,2 @@
 pub mod reformat;
+pub mod reformat_implicit_relations;
