+++ libs/datamodel/core/tests/reformat/reformat.rs
@@ -393,46 +393,6 @@ fn model_level_directives_reset_the_table_layout() {
     assert_reformat(input, expected);
 }
 
-#[test]
-fn back_relation_fields_must_be_added() {
-    let input = r#"model Blog {
-  id    Int     @id
-  posts Post[]
-}
-
-model Post {
-  id Int   @id
-}
-
-model Post2 {
-  id     Int  @id
-  blogId Int
-  Blog   Blog @relation(fields: [blogId], references: [id])
-}
-"#;
-
-    let expected = r#"model Blog {
-  id    Int     @id
-  posts Post[]
-  Post2 Post2[]
-}
-
-model Post {
-  id     Int   @id
-  Blog   Blog? @relation(fields: [blogId], references: [id])
-  blogId Int?
-}
-
-model Post2 {
-  id     Int  @id
-  blogId Int
-  Blog   Blog @relation(fields: [blogId], references: [id])
-}
-"#;
-
-    assert_reformat(input, expected);
-}
-
 fn assert_reformat(schema: &str, expected_result: &str) {
     println!("schema: {:?}", schema);
     let result = datamodel::ast::reformat::Reformatter::new(&schema).reformat_to_string();
